fix: Scope mutex holding to only what is necessary
The mutex was being held longer than necessary.
The mutex should not be held by the time a request/response
cycle is being timed as this can delay _other_ tasks
from dispatching their request. Made a small tweak to
result receivers to know when they were empty. And fixed
the elpased time variable to reflect real unit of measurement.